Moved dtos to models
